requirements.txt
labjack-ljm==1.23.0
matplotlib==3.8.3
pyserial==3.5
pyyaml==6.0.1
numpy==1.26.4